chore: clean up test infrastructure, remove dead tests
Delete tests/testcases/ and tests/customize/ (never run, no value for
milvus-lite). Convert heavy imports in common_func.py to lazy imports.
Remove unused `from utils.util_pymilvus import *` from all test files
and dead legacy fixtures from conftest. Fix numpy.NaN removal, missing
numpy alias, syntax error in search test, and argparse type bug. Add
gitignore patterns for test DB artifacts. Update devcontainer with e2e
test deps and fix CONTRIBUTING.md test instructions.

+++ tests/test_nullable.py
@@ -308,7 +308,7 @@ class TestNullableInsertAndQuery(unittest.TestCase):
         schema.add_field(
             "tag", DataType.VARCHAR, max_length=256, nullable=True, default_value="unset"
         )
-        schema.add_field("score", DataType.FLOAT, nullable=True, default_value=0.5)
+        schema.add_field("score", DataType.DOUBLE, nullable=True, default_value=0.5)
         index_params = self.client.prepare_index_params()
         index_params.add_index(field_name="vec", metric_type="L2")
         self.client.create_collection(name, schema=schema, index_params=index_params)
